--- a/prometheus.lua
+++ b/prometheus.lua
@@ -267,8 +267,7 @@ local function construct_bucket_format(buckets)
     -- floating point number with all trailing zeros removed
     local bucket_str = string.format("%f", bucket)
     local as_string = ngx_re_gsub(bucket_str, "0*$", "", "jo")
-
-    local dot_idx = utils.find(as_string,".")
+    local dot_idx = utils.find(as_string, ".")
     max_order = math.max(max_order, dot_idx - 1)
     max_precision = math.max(max_precision, as_string:len() - dot_idx)
   end