--- a/prometheus.lua
+++ b/prometheus.lua
@@ -56,9 +56,16 @@
 -- increments. Copied from https://github.com/Kong/lua-resty-counter
 local resty_counter_lib = require("prometheus_resty_counter")
 local key_index_lib = require("prometheus_keys")
-local utils = require("prometheus_utils")
+local ngx = ngx
 local ngx_re_match = ngx.re.match
 local ngx_re_gsub = ngx.re.gsub
+local error = error
+local type = type
+local get_phase = ngx.get_phase
+local ngx_sleep = ngx.sleep
+local select = select
+
+local YIELD_ITERATIONS = 200
 
 local Prometheus = {}
 local mt = { __index = Prometheus }
@@ -70,6 +77,13 @@ local TYPE_LITERAL = {
   [TYPE_COUNTER]   = "counter",
   [TYPE_GAUGE]     = "gauge",
   [TYPE_HISTOGRAM] = "histogram",
+}
+
+local can_yield_phases = {
+    rewrite = true,
+    access = true,
+    content = true,
+    timer = true
 }
 
 -- Default name for error metric incremented by this library.
@@ -153,6 +167,17 @@ local function validate_utf8_string(str)
   return true
 end
 
+-- copy form https://github.com/apache/apisix/blob/release/2.13/apisix/core/table.lua#L50-L58
+local function table_insert_tail(tab, ...)
+    local idx = #tab
+    for i = 1, select('#', ...) do
+        idx = idx + 1
+        tab[idx] = select(i, ...)
+    end
+
+    return idx
+end
+
 -- Generate full metric name that includes all labels.
 --
 -- Args:
@@ -183,7 +208,7 @@ local function full_metric_name(name, label_names, label_values)
     else
       label_value = tostring(label_values[idx])
     end
-    utils.insert_tail(label_parts, key .. '="' .. label_value .. '"')
+    table_insert_tail(label_parts, key .. '="' .. label_value .. '"')
   end
   return name .. "{" .. table.concat(label_parts, ",") .. "}"
 end
@@ -199,16 +224,15 @@ end
 --   (string) short metric name with no labels. For a `*_bucket` metric of
 --     histogram the _bucket suffix will be removed.
 local function short_metric_name(full_name)
-  local labels_start, _ = utils.find(full_name, "{")
+  local labels_start, _ = full_name:find("{")
   if not labels_start then
     return full_name
   end
   -- Try to detect if this is a histogram metric. We only check for the
   -- `_bucket` suffix here, since it alphabetically goes before other
   -- histogram suffixes (`_count` and `_sum`).
-  local suffix_idx, _ = utils.find(full_name, "_bucket{")
-
-  if suffix_idx and utils.find(full_name, "le=") then
+  local suffix_idx, _ = full_name:find("_bucket{")
+  if suffix_idx and full_name:find("le=") then
     -- this is a histogram metric
     return full_name:sub(1, suffix_idx - 1)
   end
@@ -231,7 +255,7 @@ local function check_metric_and_label_names(metric_name, label_names)
   if not metric_name:match("^[a-zA-Z_:][a-zA-Z0-9_:]*$") then
     return "Metric name '" .. metric_name .. "' is invalid"
   end
-  if utils.find(metric_name, KEY_INDEX_PREFIX) == 1 then
+  if metric_name:find(KEY_INDEX_PREFIX) == 1 then
     return "Prefix '" .. KEY_INDEX_PREFIX .. "' is reserved for internal purposes"
   end
   for _, label_name in ipairs(label_names or {}) do
@@ -267,7 +291,8 @@ local function construct_bucket_format(buckets)
     -- floating point number with all trailing zeros removed
     local bucket_str = string.format("%f", bucket)
     local as_string = ngx_re_gsub(bucket_str, "0*$", "", "jo")
-    local dot_idx = utils.find(as_string, ".")
+
+    local dot_idx = as_string:find(".", 1, true)
     max_order = math.max(max_order, dot_idx - 1)
     max_precision = math.max(max_precision, as_string:len() - dot_idx)
   end
@@ -279,11 +304,21 @@ end
 -- This function removes leading and trailing zeroes from `le` label values.
 --
 -- Args:
---   match: the result of regular matching histogram key, in the form of an array
+--   key: the metric key
 --
 -- Returns:
 --   (string) the formatted key
-local function fix_histogram_bucket_labels(match)
+local function fix_histogram_bucket_labels(key)
+  local match, err = ngx_re_match(key, METRICS_KEY_REGEX, "jo")
+  if err then
+    ngx.log(ngx.ERR, "failed to match regex: ", err)
+    return
+  end
+
+  if not match then
+    return key
+  end
+
   if match[2] == "Inf" then
     return table.concat({match[1], "+Inf", match[3]})
   else
@@ -783,6 +818,27 @@ local function register(self, name, help, label_names, buckets, typ)
   return metric
 end
 
+-- inspired by https://github.com/Kong/kong/blob/2.8.1/kong/tools/utils.lua#L1430-L1446
+-- limit to work only in rewrite, access, content and timer
+local yield
+do
+  local counter = 0
+  yield = function()
+  phase = phase or get_phase()
+    if not can_yield_phases[get_phase()] then
+      return
+    end
+
+    counter = counter + 1
+    if counter % YIELD_ITERATIONS ~= 0 then
+      return
+    end
+    counter = 0
+
+    ngx_sleep(0)
+  end
+end
+
 -- Public function to register a counter.
 function Prometheus:counter(name, help, label_names)
   return register(self, name, help, label_names, nil, TYPE_COUNTER)
@@ -820,47 +876,27 @@ function Prometheus:metric_data()
   local seen_metrics = {}
   local output = {}
   for _, key in ipairs(keys) do
-    utils.yield(true)
+    yield()
 
     local value, err = self.dict:get(key)
     if value then
-      local match, err = ngx_re_match(key, METRICS_KEY_REGEX, "jo")
-
-      if err then
-        ngx.log(ngx.ERR, "failed to match regex: ", err)
-        return
-      end
-
-      local has_seen = false
-      for seen_metric, _ in pairs(seen_metrics) do
-        if utils.has_prefix(key ,seen_metric) then
-          has_seen = true
-          break
-        end
-      end
-
-      if not has_seen then
-        local short_name = short_metric_name(key)
+      local short_name = short_metric_name(key)
+      if not seen_metrics[short_name] then
         local m = self.registry[short_name]
         if m then
           if m.help then
-            utils.insert_tail(output, string.format("# HELP %s%s %s\n",
+            table_insert_tail(output, string.format("# HELP %s%s %s\n",
             self.prefix, short_name, m.help))
           end
           if m.typ then
-            utils.insert_tail(output, string.format("# TYPE %s%s %s\n",
+            table_insert_tail(output, string.format("# TYPE %s%s %s\n",
               self.prefix, short_name, TYPE_LITERAL[m.typ]))
           end
         end
         seen_metrics[short_name] = true
       end
-
-      local fixed_key
-      if match then
-        fixed_key = fix_histogram_bucket_labels(match)
-      end
-
-      utils.insert_tail(output, string.format("%s%s %s\n", self.prefix, fixed_key or key, value))
+      key = fix_histogram_bucket_labels(key)
+      table_insert_tail(output, string.format("%s%s %s\n", self.prefix, key, value))
     else
       if type(err) == "string" then
         self:log_error("Error getting '", key, "': ", err)