The background for doing these optimizations is that when the number of keys is huge (10W+), the phenomenon of long-tail
requests is obvious.

This optimization has the following three main points

1. adjust the logic of whether the metrics are in seen_metrics
2. introduce `ngx.sleep(0)` to periodically yield the CPU and avoid long-tail requests
3. ported some optimization functions

In my test results, the latency of P100(the highest latency requests, i.e. long-tail requests) is about 1/3 ~ 1/2 of what it was before optimization

Here is the test data for the 10W keys

### before optimization

#### flamegraph
![origin-flame](https://user-images.githubusercontent.com/30819887/164509093-2972f920-832a-46fb-b83b-583ba95079b7.png)

#### wrk
```
wrk -H "Host: 1.example.com" -t4 -c100 -d30 -R5000 --u_latency http://127.0.0.1:9080/host/server/location/1/a


  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   10.21ms
 75.000%   12.03ms
 90.000%   13.25ms
 99.000%   92.80ms
 99.900%  926.72ms
 99.990%  932.86ms
 99.999%  934.40ms
100.000%  934.91ms
```

### after optimization

#### flamegraph
![origin-flame](https://user-images.githubusercontent.com/30819887/164509257-616d7a3f-201b-4f0b-b0a6-575e3f142a8e.png)

#### wrk 

```
wrk -H "Host: 1.example.com" -t4 -c100 -d30 -R5000 --u_latency http://127.0.0.1:9080/host/server/location/1/a

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%    4.34ms
 75.000%   12.81ms
 90.000%   16.12ms
 99.000%   82.75ms
 99.900%  246.91ms
 99.990%  349.44ms
 99.999%  390.40ms
100.000%  397.31ms
```