@tzssangglass 2022-04-23
Before this submission, I did the test again.

## origin master branch

```

round 1
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   76.48ms
 75.000%   88.51ms
 90.000%  105.02ms
 99.000%    1.17s
 99.900%    1.87s
 99.990%    2.03s
 99.999%    2.04s
100.000%    2.04s

round 2
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   80.51ms
 75.000%   96.13ms
 90.000%  139.90ms
 99.000%    1.22s
 99.900%    1.69s
 99.990%    1.76s
 99.999%    1.78s
100.000%    1.78s

round 3
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   90.43ms
 75.000%  103.87ms
 90.000%  140.16ms
 99.000%    1.27s
 99.900%    1.30s
 99.990%    1.40s
 99.999%    1.41s
100.000%    1.41s

round 4
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   76.54ms
 75.000%   90.05ms
 90.000%  106.50ms
 99.000%    1.16s
 99.900%    1.95s
 99.990%    1.99s
 99.999%    2.00s
100.000%    2.00s

round 5
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   80.57ms
 75.000%   87.55ms
 90.000%  106.43ms
 99.000%    1.11s
 99.900%    1.80s
 99.990%    1.84s
 99.999%    1.86s
100.000%    1.86s
```

## based on this commit, and `YIELD_ITERATIONS` is 200

```
round 1
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   89.86ms
 75.000%   99.14ms
 90.000%  139.65ms
 99.000%  315.14ms
 99.900%  838.14ms
 99.990%    1.05s
 99.999%    1.10s
100.000%    1.10s

round 2
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   81.98ms
 75.000%   92.93ms
 90.000%  126.85ms
 99.000%  211.71ms
 99.900%  723.97ms
 99.990%  939.52ms
 99.999%  958.46ms
100.000%  958.46ms

round 3
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%  101.38ms
 75.000%  109.63ms
 90.000%  127.68ms
 99.000%  239.74ms
 99.900%  927.74ms
 99.990%    1.17s
 99.999%    1.25s
100.000%    1.25s

round 4
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   91.97ms
 75.000%   98.37ms
 90.000%  116.67ms
 99.000%  370.94ms
 99.900%    1.07s
 99.990%    1.34s
 99.999%    1.37s
100.000%    1.37s

round 5
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   85.76ms
 75.000%   95.10ms
 90.000%  130.11ms
 99.000%  324.61ms
 99.900%  961.02ms
 99.990%    1.12s
 99.999%    1.19s
100.000%    1.19s
```

## based on this commit, and `YIELD_ITERATIONS` is 50

```
round 1
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   86.27ms
 75.000%   97.02ms
 90.000%  115.84ms
 99.000%  185.73ms
 99.900%  538.62ms
 99.990%  621.57ms
 99.999%  645.12ms
100.000%  645.12ms

round 2
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%   87.42ms
 75.000%   94.40ms
 90.000%  114.56ms
 99.000%  177.28ms
 99.900%  301.05ms
 99.990%  617.98ms
 99.999%  626.17ms
100.000%  626.17ms

round 3
----------------------------------------------------------

  Latency Distribution (HdrHistogram - Uncorrected Latency (measured without taking delayed starts into account))
 50.000%  106.69ms
 75.000%  115.90ms
 90.000%  133.50ms
 99.000%  214.01ms
 99.900%    1.04s
 99.990%    1.09s
 99.999%    1.13s
100.000%    1.13s
```

However, YIELD_ITERATIONS = 50 will result in a very slow response to requests for metrics that collect metrics.

From the above test results, the improvement in optimization is limited. A bit of a pity.

But I tested with wrk

```
 wrk -H "Host: 1.example.com" -t4 -c100 -d30 -R5000 --u_latency http://127.0.0.1:9080/host/server/location/1/a
```

I think in a real world scenario, it would not be a very high frequency scenario like wrk, so the main thing this optimization can do is to free up CPU time slice to handle normal requests.